Added functionality for decimals

+++ user_includes/ArrayUtilities.h
@@ -11,10 +11,10 @@
 
 #include <vector>
 
-	void printMatrix(std::vector<std::vector<int>> & matrix);
-	float det(std::vector<std::vector<int>> & matrix, int index);
-	void getCofactor(std::vector<std::vector<int>> & matrix, std::vector<std::vector<int>> & coFactors, int p, int q, int n);
-	float getCurrent(std::vector<std::vector<int>> & matrix, std::vector<int> & array, float detA, int n);
+	void printMatrix(std::vector<std::vector<double>> & matrix);
+	float det(std::vector<std::vector<double>> & matrix, int index);
+	void getCofactor(std::vector<std::vector<double>> & matrix, std::vector<std::vector<double>> & coFactors, int p, int q, int n);
+	float getCurrent(std::vector<std::vector<double>> & matrix, std::vector<double> & array, float detA, int n);
 
 
 #endif /* USER_INCLUDES_ARRAYUTILITIES_H_ */
